+++ nestedtext.py
@@ -43,8 +43,8 @@ import re
 
 
 # Globals {{{1
-__version__ = "1.3.1"
-__released__ = "2021-03-03"
+__version__ = "1.3.2"
+__released__ = "2021-04-24"
 __all__ = ['load', 'loads', 'dump', 'dumps', 'NestedTextError']